Matplotlib based plotting
Canned IPR and IRR plots using matplotlib interface.  The IPR plots are the standard, per species, plots where each process is a time series line.  The IRR plots are also per species, but each line is a reaction and the sum of reactions equals the chemistry process from the IPR plot

git-svn-id: https://dawes.sph.unc.edu/uncaqmlsvn/pyPA/trunk@652 4a91f030-2914-0410-b58d-d885616ce41a